+++ src/main/java/gregtech/common/items/MetaItems.java
@@ -325,6 +325,7 @@ public final class MetaItems {
     public static MetaItem<?>.MetaValueItem JACKHAMMER_BASE;
 
     public static MetaItem<?>.MetaValueItem NANO_SABER;
+    public static MetaItem<?>.MetaValueItem ENERGY_FIELD_PROJECTOR;
 
     public static MetaItem<?>.MetaValueItem[] DYE_ONLY_ITEMS = new MetaItem.MetaValueItem[EnumDyeColor.values().length];
     public static MetaItem<?>.MetaValueItem[] SPRAY_CAN_DYES = new MetaItem.MetaValueItem[EnumDyeColor.values().length];